refac: 재사용을 위해 useImageMetadata의 File 읽는 로직을 useImageFile로 분리
